refactor(test): move stubs and integration harness to colocated paths
Move test/stubs.ts to test/lib/stubs.ts so it lives alongside other
test helpers. Move test/lib/setup.ts to test/integration/lib/harness.ts
so the integration test harness is colocated with the integration tests
that use it.

Update all import paths in unit and integration tests accordingly.
