+++ packages/cli-core/src/test/integration/api-queries.test.ts
@@ -11,7 +11,7 @@ import {
   getInstance,
   MOCK_APP,
   MOCK_USERS,
-} from "../lib/setup.ts";
+} from "./lib/harness.ts";
 
 useIntegrationTestHarness();
