+++ packages/cli-core/src/test/integration/dry-run.test.ts
@@ -11,7 +11,7 @@ import {
   clerk,
   getInstance,
   MOCK_APP,
-} from "../lib/setup.ts";
+} from "./lib/harness.ts";
 
 useIntegrationTestHarness();
