+++ packages/cli-core/src/test/integration/env-merge.test.ts
@@ -15,7 +15,7 @@ import {
   clerk,
   getInstance,
   MOCK_APP,
-} from "../lib/setup.ts";
+} from "./lib/harness.ts";
 
 const h = useIntegrationTestHarness();
