+++ packages/cli-core/src/test/integration/onboard.test.ts
@@ -17,7 +17,7 @@ import {
   clerk,
   getInstance,
   MOCK_APP,
-} from "../lib/setup.ts";
+} from "./lib/harness.ts";
 
 const h = useIntegrationTestHarness();
