+++ packages/cli-core/src/test/lib/stubs.ts
